Day 15, late.
I think the `ToVisit.next` function is the main slow bit - since it has
to iterate over the entire map most of the time in order to find the
smallest-valued item. Adding the `last_val` parameter for a quick skip
in cases where there are multiple equal low values at some point is
good, but obviously is inconsistent.

One big performance gain in that function was changing the loop iterator
from self.map.clone() to self.map.iter() - that caused a 4x speed-up!

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>