Day 11, for now
Since the input was not too big, but did look like a pain to parse, I
put it in by hand. That didn't end up being too bad, but it would be
nice to build the monkeys from the input. But since the monkey behavior
changes between parts (because the relief no longer applies), I don't
feel like it would be necessary to print out the answers for both parts.

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>