Day 19 part one
Is it just me, or are the second parts getting harder(er)?

I wasn't sure if this "best known" thing would help very much, but it
improves the performance on the sample from about 60s to 40s! Still not
as fast as I would want for part 1 on the actual input, and definitely
not good enough for part 2.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>