Some python surprises today
I find it crazy that `.index` raises an error when stuff isn't found? If
it only runs on lists, then a `-1` sentinel value is safe... but maybe
it's overloaded for other types or something. Anyway, instead of dealing
with the error case, I guess this code goes through the lists twice,
which is kind of a bummer.

Also I have no clue what black magic that `cmp_to_key` function must
do, and I probably should've done it outside the loop...