Part two was tricky, I don't like it
Initially I thought that just checking when the guard was about to cross
their previous path (in the correct direction) would find all of the
possible spots. That finds some, but misses many, including the last one
in the examples. I was going to implement some "backtracking" - at the
very least going backwards in time for the guard to find more path
crossings, but I realized that every turn would need that same kind of
backtracking... and that seemed troublesome to track.

Thinking about the problem more before implementing that, I looked for a
trickier obstacle placement that wouldn't be covered by that algorithm.
I've included `tricky-example.txt` which shows a new obstacle that is
not on any of those paths, since it causes the guard to go in a new
direction, later interacting with the obstacle from another side and
causing a loop.

I don't know if I'm just missing something else, but brute-forcing over
all visited spots wasn't *that* terrible. The actual map might not have
any spots like my example. It would be interesting to check other
solutions against this example.