No more brute strength?
Let me tell a bit of a story here... My first approach was to memoize,
figuring that there would be a lot of repeats of the low numbers. That's
a good general trick. It was not enough, and I found that I was running
out of memory or some other limit in Python. I *should* have looked back
at the fact that "repeated numbers can be done faster" was a trick that
worked, and thought about that some more.

Instead, I thought "25? That's 5x5. I can do something with that."

So I made an `iterate5` thing, which was also memoized. Now, my program
could hit its limit about 5 (coincidence?) times faster!

I realized I only needed to *count* the number of rocks, so I could
save some memory at some stages. I also made something recursive which
would not need to keep the full state all at once. My program was no
longer hitting a limit, except for my patience.

If I had found my coding error (commented in the code), I think this
approach would've worked in a slow, but acceptable time for part 1.

Luckily, I was finally able to connect the dots and notice that since I
don't need the rocks in any order - which I had realized when I changed
to a recursive thing - I could "collapse" the rocks at each stage to
labels and counts. This allowed me to get a quick, but wrong answer
based on the sample, and debug to find my silly coder error in the
`iterate` function. And finally, part 1 and 2 were able to be completed
within 30 seconds of each other.