🎄🎄🎄
If I hadn't gotten part 2 quickly, I think I would have been very
bothered by it. My approach was to assume that the design we're looking
for was relatively compact, so if I had some measure of that for the
robots, I might be able to auto-detect it. I'm not sure if what I did
here was a rigorous statistical measure, but it worked!