day15 part2 -- I kinda liked it
I've seen this kind of double-map thing before, and it feels like it
should have a better data structure? I don't know what it would be
called. But the idea is 'stuff->id' and 'id->[stuff]'. Something built
for this would be nice because otherwise any modifications need to be
careful to keep everything synced.